modules: golang: emacs: secrets: Go modules workflow update
Sometimes Go module's root differs from project root (in terms of VCS).
Mostly lsp-mode is fine with guessing project root (thanks to Projectile
integration), but certainly not in the aforementioned case, in which it
cannot find imports, searches wrong paths and goes completely crazy.

Another option would be disabling lsp-mode root-guessing, but this does
not provide reliable experience either.

Luckily we have an option in Projectile to explicitly mark project root
with ".projectile" file, and let integration do the rest. This way we
will have nested roots in Go project in question (module's one under
VCS' one), which could seem a kind of "inaccuracy", but Gopls will be
happy.