machines: laptoptop: disabled `systemd-boot`
While migrating from legacy boot process to UEFI, it was
enabled due to respective manual (which might have been
non-precise, kinda homegrown or whatever). But since
Nixpkgs 22.05, this situation is checked and considered
conflicting. At the moment there is no observable breakages
due to its disabling, so re are moving forward this way.