emacs: history: removed `super-save`
It turns out that the package stopped working realiably, e.g. providing
its primary functionality in some circumstances, for example, high system
load - or never worked as so.

!!!
Last, but not the least - the author, Bozhidar Batsov, turns out to be
as blind and ignorant of real, unbiased situation in Ukraine, as he is
smart as a hacker. Taking this into account, I want to use little or no
emacs packages of his authorship.
!!!