appearance: emacs: `zenburn-theme` --> hc-zenburn-theme
!!!
Last, but not the least - the author, Bozhidar Batsov, turns out to be
as blind and ignorant of real, unbiased situation in Ukraine, as he is
smart as a hacker. Taking this into account, I want to use little or no
emacs packages of his authorship.
!!!

This package is *based* on the one by BB, but that is ok.