navigation: bookmarks: removed `projectile` dependency
Adding projectile's known projects in emacs externally (using [activation] script)
seemed to stop working at some time in the past. But it turns out, that this functionality
is more than fully covered with open/search modes of custom `projects` utility (from
`toolbox` project), so it is completely safe to remove it.