emacs: navigation: reverted complex `avy` setup
As always, it takes precious mental capacity to keep
track of the tool state itself, which could be used
to solve actual problems. So this is really for those
with longer/numerous "brain cache lines", or having
less tools at hand.

Nevertheless, really useful custom actions will be
added soon.