treewide: removed homebrew projects templates
There is `cachix/devenv` emerged recently, which is almost
what is wanted for development infrastructure automation.
OTOH, such templates require resources to keep them working
and up-to-date, yet not providing the wanted level of
IoC/automation. Since new `devenv` may also fall short in
this case, but being working good and also maintained, let's
stick with it and use our homebrew templates as the source of
ideas and future enligthtenment (see git diff).