wm: xmonad: switched off on-change recompilation
Unless `xmonad` enabled as nixpkgs module, there is no respective binary in PATH,
but this is not always the case. For example, we may want to generate and test
custom configuration without making `xmonad` the default WM and fully enable it.
But currently aformentioned module is the only way to actually use `xmonad`, and
there is no way to use subj in some kind of "debug mode", only enable and use at
first place. Replicating some part of this module's logic here, at configuration
level, will allow to and almost any needed workflow, so it seems as a way to go.