controlcenter: changed uptime showing implementation
Newly used `citron` tool is intended to be an on-demand status displayer,
which is not only about showing uptime. But currently it has only limited
set of measurements to display and it could be used more widely otherwise.

That said, OTOH, if we are using only its `uptime` functionality, we could
at least get rid of the hand-crafted script with quirky `sed` syntax and
the burden of it maintenance as well, though it is quite small.