git: core: emacs: added custom magit display function
It makes opening `magit-status` in new frame, when it was
issued from org-mode buffer.

Note, though, that it will reuse any newly created frame
for the very first display, because of not-yet-populated
previous buffers list for this frame (previous frame's
one will be used this time). But then, already the second
display will work as expected.