just: treewide: added task for emergency config fixup
Previously, the idea was to use vim without any customizations.
Also it was not represented as `just` task, but as `espanso` expansion
instead.

But running emacs under the same conditions is better in the sense
of not changing the editor being used, and also not installing extra
package (vim) just for the sake of emergency case. Anyway, the bare
unconfigured Emacs is good enough for this.

Also, `espanso` expansion is not cwd-dependent, thus making little to no
sense to be fired under unrelated path (other than this repo).

The case was just not thought out well beforehand. ;)