emacs: navigation: removed `phi-search-*` packages
AFAIR, they were added to make isearch-like functionality play well
with `multiple-cursors. At the moment, `isearch` seems to be used
without any interference with multiple-cursors, or at least, we are
not facing it. That said, despite the additional useful functionality
of th packages, it won't be used, as it would not be remembered, alike
with many other packages.