emacs: navigation: added prompters toggling for `embark`
When someone knows exact keybindings for various `embark` tasks,
it is more practical to issue them without any lookup (and implied
time and mind capacity loss). OTOH, when someone need or want to
discover some new tasks/keybindings, no matter the cause, it is
undoubtedly convenient to have the ability to lookup and overview
them.