modules: treewide: removed the remainder of old python scripts
Those are anyway intended to be reimplemented within `toolbox` repo [1]/[2]
likely in Go, but recently I thought of option to use Clojure/Babashka for
them, so the implementation is unclear at the moment.

Also any references to these scripts are deleted and should be (probably,
manually) resurrected/actualized from this commit, after reimplementation
is done.

Refer to [3]/[4] for the current list of removed scripts and their source
code. Following the process of reimplementation, those scripts will be
removed from [3]/[4], along with their usage context resurrection here.

[1] https://github.com/wiedzmin/toolbox
[2] [[orgit:~/workspace/repos/github.com/wiedzmin/toolbox/][~/workspace/repos/github.com/wiedzmin/toolbox/ (magit-status)]]
[3] https://github.com/wiedzmin/toolbox/tree/master/contrib
[4] [[file:~/workspace/repos/github.com/wiedzmin/toolbox/contrib]]