chassis: attributes: treewide: parameterized fill column
While editing text using monospaced font(s), there is a notion of
`fill column` which stands for column number after which the new
line of text is started, for the sake of presentation.

Here, we introduce the corresponding system-wide parameter, with
connection to underlying hardware, namely display, to keep track of
how many symbols are minimally visible on internal laptop head with
default resolution parameters (width), without any external displays.

Currently, it is being used only in Emacs, but in the future it may
become relevant elsewhere.